backend_flask/requirements.txt
alembic==1.14.0
astroid==3.3.5
blinker==1.9.0
click==8.1.7
colorama==0.4.6
dill==0.3.9
Flask==3.1.0
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
httpagentparser==1.9.5
isort==5.13.2
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.6
MarkupSafe==3.0.2
mccabe==0.7.0
platformdirs==4.3.6
PyJWT==2.10.0
pylint==3.3.1
PyMySQL==1.1.1
SQLAlchemy==2.0.36
tomlkit==0.13.2
typing_extensions==4.12.2
Werkzeug==3.1.3
flask_cors
flask-mail
flask-bcrypt
python-dotenv
cryptography
